Impliment books scrolling view in homescreen

+++ app/src/main/java/com/starry/myne/api/models/Book.kt
@@ -25,5 +25,5 @@ data class Book(
     @SerializedName("formats")
     val formats: Formats,
     @SerializedName("download_count")
-    val downloadCount: Int
+    val downloadCount: Long
 )
